Revert "Revert "hotfix: global cmdk (#2821)""
This reverts commit 178bd2bb0409e5761cb0448aeb1cd29efbb9e49f.
